requirements.txt
boto3==1.17.53
botocore==1.20.53
jmespath==0.10.0
python-dateutil==2.8.1
PyYAML==5.4.1
s3transfer==0.3.7
selenium==3.141.0
six==1.15.0
urllib3==1.26.4
